remove PLUGIN_AI_*
