--- enterprise/frontend/src/metabase-enterprise/ai-entity-analysis/analytics.ts
@@ -1,7 +1,0 @@
-import { trackSimpleEvent } from "metabase/lib/analytics";
-
-export const trackExplainChartClicked = () => {
-  trackSimpleEvent({
-    event: "metabot_explain_chart_clicked",
-  });
-};
