--- enterprise/frontend/src/metabase-enterprise/ai-entity-analysis/components/AIQuestionAnalysisSidebar/index.ts
@@ -1,1 +1,0 @@
-export { AIQuestionAnalysisSidebar } from "./AIQuestionAnalysisSidebar";
