--- enterprise/frontend/src/metabase-enterprise/ai-sql-fixer/analytics.ts
@@ -1,7 +1,0 @@
-import { trackSimpleEvent } from "metabase/lib/analytics";
-
-export const trackQueryFixClicked = () => {
-  trackSimpleEvent({
-    event: "metabot_fix_query_clicked",
-  });
-};
