+++ enterprise/frontend/src/metabase-enterprise/api/index.ts
@@ -1,4 +1,3 @@
-export * from "./ai-entity-analysis";
 export * from "./api";
 export * from "./audit-info";
 export * from "./billing-info";
