+++ enterprise/frontend/src/metabase-enterprise/plugins.ts
@@ -10,8 +10,6 @@ import "./license";
 
 // PLUGINS THAT USE hasPremiumFeature (import initialization functions):
 import { initializePlugin as initializeAdvancedPermissions } from "./advanced_permissions";
-import { initializePlugin as initializeAiEntityAnalysis } from "./ai-entity-analysis";
-import { initializePlugin as initializeAiSqlFixer } from "./ai-sql-fixer";
 import { initializePlugin as initializeApplicationPermissions } from "./application_permissions";
 import { initializePlugin as initializeAuditApp } from "./audit_app";
 import { initializePlugin as initializeAuth } from "./auth";
@@ -78,8 +76,6 @@ export function initializePlugins() {
   initializeSharing();
   initializeModeration();
   initializeAdvancedPermissions();
-  initializeAiSqlFixer();
-  initializeAiEntityAnalysis();
   initializeAuditApp();
   initializeModelPersistence();
   initializeFeatureLevelPermissions();
