+++ frontend/src/metabase/api/index.ts
@@ -51,3 +51,4 @@ export * from "./user-key-value";
 export * from "./user";
 export * from "./analytics";
 export * from "./metabot";
+export * from "./ai-entity-analysis";
