+++ frontend/src/metabase/metabot/analytics.ts
@@ -14,3 +14,15 @@ export const trackMetabotRequestSent = () => {
     event: "metabot_request_sent",
   });
 };
+
+export const trackQueryFixClicked = () => {
+  trackSimpleEvent({
+    event: "metabot_fix_query_clicked",
+  });
+};
+
+export const trackExplainChartClicked = () => {
+  trackSimpleEvent({
+    event: "metabot_explain_chart_clicked",
+  });
+};
