+++ frontend/src/metabase/metabot/components/AIAnalysisContent/AIAnalysisContent.tsx
