+++ frontend/src/metabase/metabot/components/AIAnalysisContentWrapper/AIAnalysisContentWrapper.tsx
