+++ frontend/src/metabase/metabot/components/AIDashboardAnalysisSidebar/AIDashboardAnalysisSidebar.tsx
