+++ frontend/src/metabase/metabot/components/AIQuestionAnalysisSidebar/index.ts
@@ -1,0 +1,2 @@
+export { AIQuestionAnalysisSidebar } from "./AIQuestionAnalysisSidebar";
+export type { AIQuestionAnalysisSidebarProps } from "./AIQuestionAnalysisSidebar";
