+++ frontend/src/metabase/metabot/components/AIQuestionAnalysisSidebar/utils.ts
