+++ frontend/src/metabase/metabot/components/FixSqlQueryButton/FixSqlQueryButton.tsx
