+++ frontend/src/metabase/metabot/components/FixSqlQueryButton/index.ts
