+++ frontend/src/metabase/metabot/hooks/index.ts
@@ -2,3 +2,4 @@ export { useMetabotSQLSuggestion } from "./use-metabot-sql-suggestion";
 export { useMetabotEnabledEmbeddingAware } from "./use-metabot-embedding-aware-enabled";
 export { useMetabotAgent } from "./use-metabot-agent";
 export { useMetabotAgentsManager } from "./use-metabot-agents-manager";
+export { useDashCardAnalysis } from "./useDashCardAnalysis";
