+++ frontend/src/metabase/metabot/utils/chart-analysis.ts
