+++ frontend/src/metabase/plugins/index.ts
@@ -1,13 +1,5 @@
 // Re-export all plugins from OSS modules (excluding reinitialize functions to avoid conflicts)
 export { PLUGIN_API } from "./oss/api";
-export {
-  PLUGIN_AI_SQL_FIXER,
-  PLUGIN_AI_ENTITY_ANALYSIS,
-  type PluginAiSqlFixer,
-  type AIDashboardAnalysisSidebarProps,
-  type AIQuestionAnalysisSidebarProps,
-  type PluginAIEntityAnalysis,
-} from "./oss/ai";
 export { PLUGIN_AUDIT, type InsightsLinkProps } from "./oss/audit";
 export {
   PLUGIN_AUTH_PROVIDERS,
@@ -57,10 +49,6 @@ export {
   PLUGIN_IS_EE_BUILD,
   type IllustrationValue,
 } from "./oss/core";
-export {
-  PLUGIN_DASHCARD_MENU,
-  type PluginDashcardMenu,
-} from "./oss/dashcard-menu";
 export {
   PLUGIN_DB_ROUTING,
   PLUGIN_DATABASE_REPLICATION,
@@ -163,7 +151,6 @@ export type {
 // Export a single reinitialize function that calls all individual reinitialize functions
 import { reinitialize as reinitializeNotificationsSdk } from "../../embedding-sdk-bundle/components/public/notifications";
 
-import { reinitialize as reinitializeAi } from "./oss/ai";
 import { reinitialize as reinitializeApi } from "./oss/api";
 import { reinitialize as reinitializeAudit } from "./oss/audit";
 import { reinitialize as reinitializeAuth } from "./oss/auth";
@@ -172,7 +159,6 @@ import { reinitialize as reinitializeCollections } from "./oss/collections";
 import { reinitialize as reinitializeContentTranslation } from "./oss/content-translation";
 import { reinitialize as reinitializeContentVerification } from "./oss/content-verification";
 import { reinitialize as reinitializeCore } from "./oss/core";
-import { reinitialize as reinitializeDashcardMenu } from "./oss/dashcard-menu";
 import { reinitialize as reinitializeDatabase } from "./oss/database";
 import { reinitialize as reinitializeEmbedding } from "./oss/embedding";
 import { reinitialize as reinitializeEmbeddingIframeSdk } from "./oss/embedding-iframe-sdk";
@@ -204,7 +190,6 @@ import { reinitialize as reinitializeWritableConnection } from "./oss/writable-c
 export function reinitialize() {
   reinitializeNotificationsSdk();
 
-  reinitializeAi();
   reinitializeApi();
   reinitializeAudit();
   reinitializeAuth();
@@ -213,7 +198,6 @@ export function reinitialize() {
   reinitializeContentTranslation();
   reinitializeContentVerification();
   reinitializeCore();
-  reinitializeDashcardMenu();
   reinitializeDatabase();
   reinitializeEmbedding();
   reinitializeEmbeddingIframeSdk();
