--- frontend/src/metabase/plugins/oss/ai.ts
@@ -1,65 +1,0 @@
-import type { ComponentType } from "react";
-
-import { PluginPlaceholder } from "metabase/plugins/components/PluginPlaceholder";
-import type Question from "metabase-lib/v1/Question";
-import type {
-  DashCardId,
-  Timeline,
-  TimelineEvent,
-  VisualizationDisplay,
-} from "metabase-types/api";
-
-export type PluginAiSqlFixer = {
-  FixSqlQueryButton: ComponentType<{
-    rawSql?: string | null;
-    errorMessage?: string | null;
-  }>;
-};
-
-export interface AIDashboardAnalysisSidebarProps {
-  onClose?: () => void;
-  dashcardId?: DashCardId;
-}
-
-export interface AIQuestionAnalysisSidebarProps {
-  question: Question;
-  className?: string;
-  onClose?: () => void;
-  timelines?: Timeline[];
-  visibleTimelineEvents?: TimelineEvent[];
-}
-
-export type PluginAIEntityAnalysis = {
-  AIQuestionAnalysisButton: ComponentType<any>;
-  AIQuestionAnalysisSidebar: ComponentType<AIQuestionAnalysisSidebarProps>;
-  AIDashboardAnalysisSidebar: ComponentType<AIDashboardAnalysisSidebarProps>;
-  canAnalyzeQuestion: (question: Question) => boolean;
-  chartAnalysisRenderFormats: {
-    [display in VisualizationDisplay]?: "png" | "svg" | "none";
-  };
-};
-
-const getDefaultPluginAiSqlFixer = (): PluginAiSqlFixer => ({
-  FixSqlQueryButton: PluginPlaceholder,
-});
-export const PLUGIN_AI_SQL_FIXER: PluginAiSqlFixer =
-  getDefaultPluginAiSqlFixer();
-
-const getDefaultPluginAIEntityAnalysis = (): PluginAIEntityAnalysis => ({
-  AIQuestionAnalysisButton: PluginPlaceholder,
-  AIQuestionAnalysisSidebar: PluginPlaceholder,
-  AIDashboardAnalysisSidebar: PluginPlaceholder,
-  canAnalyzeQuestion: () => false,
-  chartAnalysisRenderFormats: {},
-});
-
-export const PLUGIN_AI_ENTITY_ANALYSIS: PluginAIEntityAnalysis =
-  getDefaultPluginAIEntityAnalysis();
-
-/**
- * @internal Do not call directly. Use the main reinitialize function from metabase/plugins instead.
- */
-export function reinitialize() {
-  Object.assign(PLUGIN_AI_SQL_FIXER, getDefaultPluginAiSqlFixer());
-  Object.assign(PLUGIN_AI_ENTITY_ANALYSIS, getDefaultPluginAIEntityAnalysis());
-}
