+++ frontend/src/metabase/query_builder/components/view/View/NativeQueryRightSidebar/NativeQueryRightSidebar.tsx
@@ -1,6 +1,6 @@
 import { match } from "ts-pattern";
 
-import { PLUGIN_AI_ENTITY_ANALYSIS } from "metabase/plugins";
+import { AIQuestionAnalysisSidebar } from "metabase/metabot/components/AIQuestionAnalysisSidebar";
 import type { EmbeddingParameterVisibility } from "metabase/public/lib/types";
 import { TagEditorSidebar } from "metabase/query_builder/components/template_tags/TagEditorSidebar";
 import { QuestionInfoSidebar } from "metabase/query_builder/components/view/sidebars/QuestionInfoSidebar";
@@ -126,7 +126,7 @@ export const NativeQueryRightSidebar = (
       <QuestionSettingsSidebar question={question} />
     ))
     .with({ isShowingAIQuestionAnalysisSidebar: true }, () => (
-      <PLUGIN_AI_ENTITY_ANALYSIS.AIQuestionAnalysisSidebar
+      <AIQuestionAnalysisSidebar
         question={question}
         visibleTimelineEvents={timelineEvents}
         timelines={timelines}
