+++ frontend/src/metabase/query_builder/containers/use-register-query-builder-metabot-context.tsx
@@ -3,7 +3,7 @@ import { match } from "ts-pattern";
 
 import { useRegisterMetabotContextProvider } from "metabase/metabot";
 import { useMetabotEnabledEmbeddingAware } from "metabase/metabot/hooks";
-import { PLUGIN_AI_ENTITY_ANALYSIS } from "metabase/plugins";
+import { CHART_ANALYSIS_RENDER_FORMATS } from "metabase/metabot/utils/chart-analysis";
 import {
   getChartImagePngDataUri,
   getChartSelector,
@@ -161,8 +161,9 @@ function getVisualizationDataUri(question: Question) {
   const display = question.card().display;
 
   const format =
-    PLUGIN_AI_ENTITY_ANALYSIS.chartAnalysisRenderFormats[display] ??
-    ("none" as const);
+    (CHART_ANALYSIS_RENDER_FORMATS as Record<string, "png" | "svg" | "none">)[
+      display
+    ] ?? ("none" as const);
 
   return match(format)
     .with("none", () => undefined)
