+++ frontend/src/metabase/query_builder/containers/use-register-query-builder-metabot-context.unit.spec.tsx
@@ -1,6 +1,5 @@
 import _ from "underscore";
 
-import { setupEnterpriseOnlyPlugin } from "__support__/enterprise";
 import { mockSettings } from "__support__/settings";
 import { renderHookWithProviders } from "__support__/ui";
 import type { ComputedVisualizationSettings } from "metabase/visualizations/types";
@@ -16,7 +15,6 @@ import {
   createMockSettings,
   createMockSingleSeries,
   createMockTimelineEvent,
-  createMockTokenFeatures,
   createMockVisualizationSettings,
 } from "metabase-types/api/mocks";
 import { createAdHocCard } from "metabase-types/api/mocks/presets";
@@ -93,18 +91,6 @@ function createMockData(opts: {
 }
 
 describe("registerQueryBuilderMetabotContextFn", () => {
-  beforeAll(() => {
-    mockSettings(
-      createMockSettings({
-        "token-features": createMockTokenFeatures({
-          ai_entity_analysis: true,
-        }),
-      }),
-    );
-    setupEnterpriseOnlyPlugin("metabot");
-    setupEnterpriseOnlyPlugin("ai-entity-analysis");
-  });
-
   beforeEach(() => {
     jest.clearAllMocks();
   });
