+++ frontend/src/metabase/querying/components/CodeMirrorEditor/util.unit.spec.tsx
@@ -1,7 +1,5 @@
 import { EditorState } from "@codemirror/state";
 
-import { setupEnterpriseOnlyPlugin } from "__support__/enterprise";
-
 import { getPlaceholderText, matchTagAtCursor } from "./util";
 
 describe("matchTagAtCursor", () => {
@@ -174,10 +172,6 @@ describe("getPlaceholderText", () => {
   });
 
   describe("metabot is enabled", () => {
-    beforeAll(() => {
-      setupEnterpriseOnlyPlugin("metabot");
-    });
-
     it("should return metabot placeholder text for sql", () => {
       expect(getPlaceholderText("sql", true)).toBe(
         "Write your SQL here, or press Ctrl + Shift + i to have SQL generated for you.",
