+++ frontend/test/__support__/enterprise-typed.ts
@@ -13,13 +13,11 @@ export const ENTERPRISE_PLUGIN_NAME_MAPPING = {
   sharing: "metabase-enterprise/sharing",
   metabot: "metabase/metabot",
   rich_text_editing: "metabase-enterprise/rich_text_editing",
-  "ai-entity-analysis": "metabase-enterprise/ai-entity-analysis",
   license: "metabase-enterprise/license",
   user_provisioning: "metabase-enterprise/user_provisioning",
   application_permissions: "metabase-enterprise/application_permissions",
   auth: "metabase-enterprise/auth",
   "smtp-override": "metabase-enterprise/smtp-override",
-  "ai-sql-fixer": "metabase-enterprise/ai-sql-fixer",
   embedding: "metabase-enterprise/embedding",
   admin: "metabase-enterprise/admin",
   dependencies: "metabase-enterprise/dependencies",
